refactor: Config 폴더계층 수정
