+++ src/main/java/OrangeCorps/LBridge/Service/News/NewsService.java
@@ -1,7 +1,7 @@
 package OrangeCorps.LBridge.Service.News;
 
-import static OrangeCorps.LBridge.Config.My_API_KEY_IS;
-import static OrangeCorps.LBridge.Config.QUERY_IS;
+import static OrangeCorps.LBridge.Config.Config.My_API_KEY_IS;
+import static OrangeCorps.LBridge.Config.Config.QUERY_IS;
 
 import OrangeCorps.LBridge.Domain.News.NewsDTO;
 import OrangeCorps.LBridge.Service.CoupleService;
@@ -11,7 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
-import static OrangeCorps.LBridge.Config.*;
+import static OrangeCorps.LBridge.Config.Config.*;
 
 @Service
 public class NewsService {
